pyproject.toml
[project]
name = "img-db"
version = "0.3.0"
description = "img-DB is a CLI application for managing your images"
authors = [
    { name = "Cristi Constantin" }
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["image", "photo", "database"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Topic :: Multimedia :: Graphics"
]

packages = [
    { include = "imgdb" },
    { exclude = "notebook" },
]

requires-python = "~= 3.13"

dependencies = [
    "pillow ~= 11.1",
    "numpy ~= 2.2",
    "scipy ~= 1.15",
    "lxml ~= 5.3",
    "beautifulsoup4 ~= 4.12",
    "blurhash-python ~= 1.2",
    "texttable ~= 1.7",
    "tqdm ~= 4.67",
    "attrs ~= 24.3",
    "Jinja2 ~= 3.1",
]

[project.urls]
Homepage = "https://github.com/croqaz/img-db"
Repository = "https://github.com/croqaz/img-db"
Documentation = "https://github.com/croqaz/img-db/tree/main/docs"
Issues = "https://github.com/croqaz/img-db/issues"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[behavior]
use-uv = true
autosync = true

[dependency-groups]
dev = [
    "hatch ~= 1.14",
    "ruff ~= 0.8",
    "mypy ~= 1.14",
    "pytest ~= 8.3",
    "pytest-cov ~= 6.0",
    "codecov ~= 2.1",
    "ipython ~= 8.31",
    "q ~= 2.7",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["imgdb"]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C",  # flake8-comprehensions
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "SIM", # flake8-simplify
    "RUF100", # unused-noqa
]
ignore = ["E731", "SIM109", "C901"]

[tool.ruff.lint.per-file-ignores]
"imgdb/__init__.py" = ["F"]

[tool.coverage.report]
ignore_errors = true
omit = [
    "imgdb/chart.py",
]
exclude_lines = [
    "def __repr__",
    "def __str__",
    "if __name__ == __main__:",
    "pragma: no cover",
    "warnings\\.warn.+",
]
